[Fix] デバッグコマンドで元素使いに職業を変更して魔法を使用するとクラッシュする
元素使いの魔法領域は通常の領域と別扱いなのでデバッグコマンドで領域選択が
行われず正しい領域が選択されていない状態のままとなるので魔法を使用すると
クラッシュする。
元素使い用の領域選択処理を実装し職業変更時に領域を選択できるようにする。
また領域変更コマンドにも対応する。

+++ src/birth/quick-start.cpp
@@ -89,7 +89,7 @@ void save_prev_data(PlayerType *player_ptr, birther *birther_ptr)
     birther_ptr->ppersonality = player_ptr->ppersonality;
 
     if (PlayerClass(player_ptr).equals(PlayerClassType::ELEMENTALIST)) {
-        birther_ptr->realm1 = player_ptr->element;
+        birther_ptr->realm1 = static_cast<int16_t>(player_ptr->element);
         birther_ptr->realm2 = 0;
     } else {
         PlayerRealm pr(player_ptr);
